server/requirements.txt
Flask
pymongo
python-dotenv
flask_pymongo